Create hide-4k-TV.py
This script will iterate through all TV shows in a given library, then for each show, iterate through all episodes. If a show has an episode which is only available in 4K, it will mark the show as "Untranscodable". If all episodes of a show are transcodable, it will remove the label.